chore: move the code to tagsrepository
This may be a problem later on, but I want to try it, because it removes many logic from posts that belongs to tags.
